+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 # 0.3.20-SNAPSHOT
 
+- adjustment width - stopped using `w-1/n`.
+- reverse order tasks/answers, comments, stocks.
 - hkim0331/konpy2:0.3.19 is amd64 binary. use `konpy2:latest`
 - updated `deps.edn` for docker container - added aliases :dev and :nrepl
 - fixed: did not forward 5555 from container - hkimjp.konpy2.* was not found.